img and location

+++ src/app/tab2/tab2.page.ts
@@ -1,4 +1,4 @@
-import { Component, ViewChild, inject } from '@angular/core';
+import { Component, OnDestroy, ViewChild, inject } from '@angular/core';
 import { ExploreContainerComponent } from '../explore-container/explore-container.component';
 import { NoteService } from '../services/note.service';
 import { CommonModule } from '@angular/common';
@@ -6,6 +6,9 @@ import { IonList, IonicModule } from '@ionic/angular';
 import { Note } from '../model/note';
 import { ModalController } from '@ionic/angular';
 import { EditModalComponent } from '../components/edit-modal/edit-modal.component';
+import { UIService } from '../services/ui.service';
+import { Subscription } from 'rxjs';
+import { NoteModalComponent } from '../components/note-modal/note-modal.component';
 
 @Component({
   selector: 'app-tab2',
@@ -14,52 +17,78 @@ import { EditModalComponent } from '../components/edit-modal/edit-modal.componen
   standalone: true,
   imports: [IonicModule, ExploreContainerComponent, CommonModule]
 })
-export class Tab2Page {
+export class Tab2Page implements OnDestroy {
   public noteS = inject(NoteService);
   public modalController = inject(ModalController);
+  private UIS = inject(UIService);
+  private suscription!:Subscription;
+
+  public notesList:Note[] = [];
 
   @ViewChild('lista') lista!: IonList;
 
   constructor() { }
 
   ionViewDidEnter() {
-    this.noteS.readAll()
+    this.suscription = this.noteS.readAll().subscribe(value => {
+      if(value){
+        this.notesList = value
+      }  
+    })
+  }
+
+  ngOnDestroy(){
+    this.suscription.unsubscribe();
   }
 
   editNote() {
 
   }
 
-  deleteNote(note: Note) {
-    try {
-      this.noteS.deleteNote(note);
+  async deleteNote(note: Note) {
+    if(await this.UIS.confirmation()==='confirm'){
+      try {
+        this.noteS.deleteNote(note);
+        this.lista.closeSlidingItems();
+      } catch {
+  
+      }
+    }else{
       this.lista.closeSlidingItems();
-    } catch {
-
     }
+    
   }
 
   onItemSlide(event: any, note: Note) {
     const swipeDirection = event.detail.side;
 
     if (swipeDirection === "start") {
-      this.openModal(note);
+      this.openModal(note, EditModalComponent);
       // this.editNote();
     } else if (swipeDirection === "end") {
       this.deleteNote(note);
     }
   }
 
-  async openModal(note: Note) {
-    const modal = await this.modalController.create({
-      component: EditModalComponent,
-      componentProps: {
-        // Puedes pasar propiedades al modal si es necesario
-        nota: note
-      },
-      cssClass: 'editModal',
-    });
+  async openModal(note: Note, modalSet:any) {
+    this.UIS.openModal(note, modalSet);
     this.lista.closeSlidingItems();
-    await modal.present();
+  }
+
+  // async openModal(note: Note, modalSet:any) {
+  //   const modal = await this.modalController.create({
+  //     component: EditModalComponent,
+  //     componentProps: {
+  //       // Puedes pasar propiedades al modal si es necesario
+  //       nota: note
+  //     },
+  //     cssClass: 'editModal',
+  //   });
+  //   this.lista.closeSlidingItems();
+  //   await modal.present();
+  // }
+
+  viewNote(note:Note){
+    this.openModal(note, NoteModalComponent);
   }
 }
